JokeService JSON parsing test now passes
