+++ CanHazFunny/CanHazFunny.Tests/JesterTests.cs
@@ -63,20 +63,5 @@ namespace CanHazFunny.Tests;
             outputMock.Verify(OutputToScreen => OutputToScreen.WriteJokeToScreen(noChuckJoke), Times.Once());
         }
 
-        [Fact]
-        public void ParseJokeJSON_ParseJSon_ReturnsJokeAsString()
-        {
-            string jokeJSON = "{{\"joke\": \"Jesus can walk on water, but Chuck Norris can walk on Jesus.\"}}";
-            string expectedParsedString = "Jesus can walk on water, but Chuck Norris can walk on Jesus.";
-
-            string actualParsedString = Jester.ParseJokeJSON(jokeJSON);
-
-            // Mark said this was okay to use generic equals method for strings
-            #pragma warning disable xUnit2006 // Do not use invalid string equality check
-            Assert.Equal<string>(expectedParsedString, actualParsedString);
-            #pragma warning restore xUnit2006 // Do not use invalid string equality check
-
-    }
-
     }
