+++ CanHazFunny/CanHazFunny/JokeService.cs
@@ -3,10 +3,11 @@ using System.Net.Http;
 using System.Net.Http.Json;
 using System.Text.Json;
 using System.Text.Json.Serialization;
+using System.Text.RegularExpressions;
 
 namespace CanHazFunny;
 
-    public class JokeService : IJokeService
+    public partial class JokeService : IJokeService
     {
         private HttpClient HttpClient { get; } = new();
         public string GetJoke()
@@ -17,4 +18,17 @@ namespace CanHazFunny;
 
             return joke;
         }
-    }
+
+        public static string ParseJokeJSON(string jokeJSON)
+        {
+            int beginningIndex = jokeJSON.IndexOf(": \"") + 3;
+            int ending = jokeJSON.IndexOf("\"}}");
+
+
+
+            return jokeJSON[beginningIndex..ending];
+            
+
+        }
+
+}
